Add new property 'profile.uvp.document.store.base.url' to be able to set a base url for the external access to the document store.
This holds the base url that points to the document store. It will be
used as a base for generating the download links in the uvp portal
detail view. This property is specific to the uvp profile.

The property was introduced to be able to connect UVP NI installation 
to uvp-verbund.de. Since all document urls were relative before. With
this property set to an absolute base url, exchange of datasets between
uvp ingrid installations is no problem anymore. 

see also https://redmine.informationgrid.eu/issues/915