Allow CSW-T import under folders (#2088)
When importing an XML document using CSW-T, it is checked whether the
parent document (if set in the XML document) has been published or not.
For folders this check always fails because folders cannot be published.
Disable checks for the publication status of folders to allow import to
work when the parentIdentifier refers to an existing folder.