Add IDF mapper that handles metadata date. #1084
IDF Mapper to adjust the metadata date (gmd:dateStamp) according to the
changes to the ISO 19139 XML format compared to the previous export.

Therefore a fingerprint is taken from the ISO 19139 XML format and stored
in the database. The Mapper compares the generated ISO 19139 XML format
with the stored fingerprint. If the fingerprints differ, the metadata date
is changed to the current date.

See https://redmine.informationgrid.eu/issues/1084 for details.